requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.24
python-multipart==0.0.6
cryptography==41.0.7
python-jose[cryptography]==3.3.0
pytest==7.4.3
httpx==0.25.2
motor==3.7.1
pymongo==4.15.4
python-dotenv==1.2.1